moved data files for frontend to utils folder
