--- data/planning_time.csv
@@ -1,43 +1,0 @@
-year,week,Duration
-2023,51,21.0
-2023,52,43.0
-2024,1,77.0
-2024,2,32.0
-2024,3,46.0
-2024,4,120.0
-2024,5,136.0
-2024,6,42.0
-2024,7,21.0
-2024,8,41.0
-2024,10,126.0
-2024,13,25.0
-2024,15,199.0
-2024,16,216.0
-2024,17,177.0
-2024,21,50.0
-2024,22,36.0
-2024,23,52.0
-2024,24,97.0
-2024,26,9.0
-2024,36,8.0
-2024,38,41.0
-2024,39,62.0
-2024,40,65.0
-2024,41,78.0
-2024,42,21.0
-2024,43,12.0
-2024,44,18.0
-2024,45,40.0
-2024,47,67.0
-2024,48,15.0
-2024,49,141.0
-2024,51,81.0
-2024,52,13.0
-2025,2,34.0
-2025,3,56.0
-2025,4,88.0
-2025,5,132.0
-2025,6,61.0
-2025,7,103.0
-2025,8,50.0
-2025,9,6.0
