+++ src/frontend/src/components/StatCardSystem/StatCardSystem.jsx
@@ -1,0 +1,12 @@
+import { useContext, useMemo } from 'react';
+import { TaskContext } from '@/contexts/TaskContext';
+
+function StatCardSystem() {
+    const { tasks, timeSum} = useContext(TaskContext);
+
+    return (
+        <div>{timeSum}</div>
+    );
+}
+
+export default StatCardSystem;
