+++ src/frontend/src/utils/categories.json
