+++ src/frontend/src/utils/computeStatuses.js
@@ -1,0 +1,5 @@
+const computeStatuses = () => {
+    
+}
+
+export default computeStatuses;
