+++ src/frontend/src/utils/default_tasks.jsx
